SKALE-1373 build fix with clang

+++ datastructures/CommittedBlockList.h
@@ -38,7 +38,7 @@ class CommittedBlock;
 class CommittedBlockList : public DataStructure {
     ptr< vector< ptr< CommittedBlock > > > blocks = nullptr;
 
-    CommittedBlockList( ptr< vector< size_t > > _blockSizes,
+    CommittedBlockList( ptr< vector< uint64_t > > _blockSizes,
         ptr< vector< uint8_t > > _serializedBlocks, uint64_t offset = 0 );
 
 
@@ -48,17 +48,14 @@ public:
 
     ptr< vector< ptr< CommittedBlock > > > getBlocks();
 
-    ptr<vector<uint64_t >> createSizes();
+    ptr< vector< uint64_t > > createSizes();
 
     ptr< vector< uint8_t > > serialize();
 
-    static ptr< CommittedBlockList > deserialize(ptr<vector<size_t>> _blockSizes,
-                                                              ptr< vector< uint8_t > > _serializedBlocks,
-                                                              uint64_t _offset);
+    static ptr< CommittedBlockList > deserialize( ptr< vector< uint64_t > > _blockSizes,
+        ptr< vector< uint8_t > > _serializedBlocks, uint64_t _offset );
 
 
     static ptr< CommittedBlockList > createRandomSample( uint64_t _size,
         boost::random::mt19937& _gen, boost::random::uniform_int_distribution<>& _ubyte );
-
-
 };
